+++ lib/screens/roles/common.dart
@@ -1,0 +1,57 @@
+import 'package:flutter/material.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+
+Future<void> logout(BuildContext context) async {
+  final prefs = await SharedPreferences.getInstance();
+  await prefs.remove('access_token');
+  await prefs.remove('role');
+  await prefs.remove('id');
+  await prefs.remove('full_name');
+  if (context.mounted) {
+    Navigator.of(context).pushNamedAndRemoveUntil('/login', (route) => false);
+  }
+}
+
+PreferredSizeWidget roleAppBar(String title, BuildContext context) {
+  return AppBar(
+    title: Text(title),
+    actions: [
+      IconButton(
+        onPressed: () => logout(context),
+        icon: const Icon(Icons.logout),
+        tooltip: 'Logout',
+      ),
+    ],
+  );
+}
+
+class StatCard extends StatelessWidget {
+  final String label;
+  final String value;
+  final IconData icon;
+  final Color color;
+
+  const StatCard({super.key, required this.label, required this.value, required this.icon, required this.color});
+
+  @override
+  Widget build(BuildContext context) {
+    return Card(
+      elevation: 2,
+      child: Padding(
+        padding: const EdgeInsets.all(16),
+        child: Row(
+          children: [
+            CircleAvatar(backgroundColor: color.withOpacity(0.15), child: Icon(icon, color: color)),
+            const SizedBox(width: 12),
+            Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
+              Text(value, style: Theme.of(context).textTheme.titleLarge?.copyWith(fontWeight: FontWeight.bold)),
+              const SizedBox(height: 4),
+              Text(label, style: Theme.of(context).textTheme.bodySmall),
+            ]),
+          ],
+        ),
+      ),
+    );
+  }
+}
+
